pyproject.toml
[project]
name = "agent-gaia"
version = "0.1.0"
description = "Multi-LLM RFP Analysis Platform with 3-way WebSocket streaming"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0",
    "hvac>=2.3.0",
    "tenacity>=8.2.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.12",
    "jinja2>=3.1.5",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
]
